% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getip-package.r
\docType{package}
\name{getip-package}
\alias{getip-package}
\title{getip}
\description{
A micro-package for getting your IP address, either the
local or external one.
}
\author{
Drew Schmidt
}
\keyword{package}

