% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getip.r
\name{getip}
\alias{getip}
\title{ip_external}
\usage{
getip(type = "internal")
}
\value{
Returns the local IP address as a string.
}
\description{
Get your internal/local or external/public IP address.
}
\details{
The lookup is done in-process, i.e., it does not call
\code{system()} or the like.

The function returns a lookup from one of several services, including
Amazon AWS \url{http://checkip.amazonaws.com/}, 
httpbin \url{http://httpbin.org/ip}, ipify \url{https://www.ipify.org/},
and "My External IP address is ..." \url{http://myexternalip.com/}.
You must be connected to the internet for this to work.
}
\examples{
\dontrun{
# internal/local address
getip("internal")

# external/public
getip("external")
}

}

