% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getip.r
\name{ip_external}
\alias{ip_external}
\title{ip_internal}
\usage{
ip_external()
}
\value{
Returns the local IP address as a string.
}
\description{
Get the local/internal IP address.
}
\details{
The function returns a lookup from \url{http://httpbin.org/ip}.
You must be connected to the internet for this to work.
}
\examples{
\dontrun{
ip_external()
}

}

