% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_internal.r
\name{ip_internal}
\alias{ip_internal}
\title{ip_internal}
\usage{
ip_internal()
}
\value{
Returns the local IP address as a string.
}
\description{
Get your local/internal IP address.
}
\details{
The lookup is done in-process, i.e., it does not call
\code{system()} or the like.
}
\examples{
\dontrun{
ip_internal()
}

}

