% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validip.r
\name{validip}
\alias{validip}
\title{validip}
\usage{
validip(ip)
}
\arguments{
\item{ip}{A string.}
}
\value{
Returns \code{TRUE} if the input is a potentially valid IP address, and
\code{FALSE} otherwise.
}
\description{
Check if a string is a valid IP address. Currently only IPv4 addresses are
supported.
}
\details{
This function is not like a 'ping'. It merely checks if the string is
correctly formatted and could theoretically be a valid IP address.
}
\examples{
library(getip)

validip("192.168.1.1")
validip("700.168.1.1")

}
