% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getip-package.r
\docType{package}
\name{getip-package}
\alias{getip-package}
\title{getip}
\description{
A micro-package for getting your IP address, either the local/internal or the
public/external one.  Currently only IPv4 addresses are supported.
}
\author{
Drew Schmidt
}
\keyword{package}
