% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ip_external.r
\name{ip_external}
\alias{ip_external}
\title{ip_external}
\usage{
ip_external()
}
\value{
Returns the local IP address as a string.
}
\description{
Get your external/public IP address.
}
\details{
The function returns a lookup from one of several services, including
Amazon AWS \url{http://checkip.amazonaws.com/}, 
httpbin \url{http://httpbin.org/ip}, ipify \url{https://www.ipify.org/},
and "My External IP address is ..." \url{http://myexternalip.com/}.
You must be connected to the internet for this to work.
}
\examples{
\dontrun{
ip_external()
}

}

